Quitando warnings

+++ PGUPV/include/window.h
@@ -96,7 +96,7 @@ namespace PGUPV {
 		void setTitle(const std::string &title);
 		// Mueve el puntero del ratón a las coordenadas indicadas (cuidado!
 		// genera el evento de movimiento de ratón correspondiente)
-		void setMousePosition(uint x, uint y);
+		void setMousePosition(float x, float y);
 		// Muestra/oculta el puntero del ratón
 		void showMouseCursor(bool show = true);
 		/**
@@ -247,7 +247,7 @@ namespace PGUPV {
 		JoystickButtonEventsSource::SubscriptionId joyButtonSubsId;
 
 		std::weak_ptr<Renderer> activeRenderer;
-		uint lastXMousePos, lastYMousePos;
+		float lastXMousePos, lastYMousePos;
 
 		void buildStatsPanel();
 	};
